pyproject.toml
[project]
name = "bliq"
version = "0.1.0"
description = "Binance USDT-M perpetual futures liquidity measurement CLI"
requires-python = ">=3.11"
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "typer>=0.12",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "loguru>=0.7",
    "rich>=13.7",
    "websockets>=13.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
]

[project.scripts]
bliq = "bliq.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bliq"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: tests that hit real Binance endpoints (opt-in)",
]
addopts = "-ra --strict-markers"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.format]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["N802", "N803"]
